Format Letters And Arrows Puzzle
Set RandomSeed in the contructor. Remove redundant tests for formatting. Add a self-correcting unit test for formatting. This test will always pass on the second run, so it's of limited usefulness. Hopefully, I'll see the difference in the expected files when I check in.
Added todo list. All tests pass.